refactor(logistics): extract field visit id normalization
